Quotient preliminaries cleaned up
Added more stuff to quotient preliminaries (what it means to quotient a
graph by a group) and moved it earlier.  Probably won’t matter later
anyway when we delete a lot of the stuff that is standard, but it helps
to have it here for now to make sure we have everything defined and
logically correct.