New alternative proof in altproof
New file altproof.tex that gives an alternative proof of the main
theorem.  It doesn’t use I^n/C like the other proof, and it is more in
keeping with a graph theoretic approach.  Still very rough but it looks
nicer.  I didn’t add it to the main file because I’m not sure yet if we
want to do this.