ESDE codes earlier, more fleshed out
Instead of a later section of ESDE codes, I have a section that
explains the intuition (with examples) of each of the 3 pieces of the
ESDE code (the code for A_L^0, the code for A_R^0, and K).   This is
all before the proof of the main theorem.  Classifying ESDE codes is
also placed here.

In addition, I am using notation A’ = A_L^0 x A_R^0, and its
corresponding code C’, throughout, for brevity and readability
(especially now that we have commutative diagrams where space is
valuable).

Unfortunately it does break up the flow to the proof of the main
theorem but we introduce the main theorem in the introduction anyway,
and some of the ideas here are useful to understanding the proof.  This
also solves a few problems that were mentioned in comments earlier,
about introducing A_L and A_R before we use it, and commenting about K
and what it really is.

In addition, I moved the corollaries to the main theorem (structure
theorems about 2d Adinkras) to a new section after the proof.  That
allows the proof (which may not be of interest to the reader) to be
isolated in one section, and makes it easier to find the corollaries.
Enumerating 2d Adinkras is now a subsection of this new section.