Cleaned up some \pi_L/R use
Removed most of the uses of \pi_L and \pi_R except for the construction
of 2d Adinkras using the main theorem; this remains useful.  Other uses
turned out to be better explained using x_L and x_R, like the proof
that \Psi: K -> V^0 is a bijection.  In definition of ESDE codes, we
use wt_L and wt_R because it’s more intuitive and it’s used later
anyway.