Change figure styles
Changed the figure styles to be more like Adinkras in older papers (but
no distinguishing bosons from fermions).  Note that nodes are not
labeled and are small dots.  Reduced the number of 1d example Adinkras
to one.  Removed the example of the Valise 2-d Adinkra and simply cited
the one from the figure for 2d Adinkras.