figure for vertex switching, reorder citations, tweak other figure
Added a new figure for vertex switching
Reordered citations when they can go in numerical order (except when
possible, cite the original Adinkra paper first).
In the quotient figure, change the order of painting the edges so that
lines in front look like they are in front (not an escher drawing).